+++ tests/Paket.Bootstrapper.Tests/DownloadStrategies/CacheDownloadStrategyTest.cs
@@ -135,10 +135,10 @@ namespace Paket.Bootstrapper.Tests.DownloadStrategies
             mockFileProxy.Setup(x => x.FileExists(ItHasFilename("paket.exe"))).Returns(true);
 
             //act
-            sut.DownloadVersion("any", "any");
+            sut.DownloadVersion("any", "any", null);
 
             //assert
-            mockEffectiveStrategy.Verify(x => x.DownloadVersion(It.IsAny<string>(), It.IsAny<string>()), Times.Never);
+            mockEffectiveStrategy.Verify(x => x.DownloadVersion(It.IsAny<string>(), It.IsAny<string>(), null), Times.Never);
             mockFileProxy.Verify(x => x.CopyFile(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<bool>()));
         }
 
@@ -149,10 +149,10 @@ namespace Paket.Bootstrapper.Tests.DownloadStrategies
             mockFileProxy.Setup(x => x.FileExists(It.IsAny<string>())).Returns(false);
 
             //act
-            sut.DownloadVersion("any", "any");
+            sut.DownloadVersion("any", "any", null);
 
             //assert
-            mockEffectiveStrategy.Verify(x => x.DownloadVersion("any", "any"));
+            mockEffectiveStrategy.Verify(x => x.DownloadVersion("any", "any", null));
             mockFileProxy.Verify(x => x.CopyFile(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<bool>()));
         }
 
@@ -170,23 +170,27 @@ namespace Paket.Bootstrapper.Tests.DownloadStrategies
         [Test]
         public void GetLatestVersion_PaketFileCorrupt_DownloadPaketFile()
         {
+            const string hashFile = @"C:\hash.txt";
+
             //arrange
             mockEffectiveStrategy.Setup(x => x.GetLatestVersion(true)).Throws<WebException>().Verifiable();
             mockFileProxy.Setup(x => x.OpenRead(It.IsAny<string>())).Returns(new MemoryStream(Guid.NewGuid().ToByteArray()));
-            mockFileProxy.Setup(x => x.ReadAllLines(It.IsAny<string>())).Returns(new[] { Guid.NewGuid().ToString().Replace("-", String.Empty) + " paket.exe" });
+            mockFileProxy.Setup(x => x.ReadAllLines(hashFile)).Returns(new[] { Guid.NewGuid().ToString().Replace("-", String.Empty) + " paket.exe" });
             mockFileProxy.Setup(x => x.FileExists(It.IsAny<string>())).Returns(true);
 
             //act
-            sut.DownloadVersion("any", "any");
+            sut.DownloadVersion("any", "any", hashFile);
 
             //assert
-            mockEffectiveStrategy.Verify(x => x.DownloadVersion(It.IsAny<string>(), It.IsAny<string>()), Times.Once);
+            mockEffectiveStrategy.Verify(x => x.DownloadVersion(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>()), Times.Once);
             mockFileProxy.Verify(x => x.CopyFile(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<bool>()));
         }
 
         [Test]
         public void GetLatestVersion_PaketFileNotCorrupt_DontDownloadPaketFile()
         {
+            const string hashFile = @"C:\hash.txt";
+
             //arrange
             var content = Guid.NewGuid().ToByteArray();
             var sha = new SHA256Managed();
@@ -194,37 +198,87 @@ namespace Paket.Bootstrapper.Tests.DownloadStrategies
             var hash = BitConverter.ToString(checksum).Replace("-", String.Empty);
             mockEffectiveStrategy.Setup(x => x.GetLatestVersion(true)).Throws<WebException>().Verifiable();
             mockFileProxy.Setup(x => x.OpenRead(It.IsAny<string>())).Returns(new MemoryStream(content));
-            mockFileProxy.Setup(x => x.ReadAllLines(It.IsAny<string>())).Returns(new[] { hash + " paket.exe" });
+            mockFileProxy.Setup(x => x.ReadAllLines(hashFile)).Returns(new[] { hash + " paket.exe" });
             mockFileProxy.Setup(x => x.FileExists(It.IsAny<string>())).Returns(true);
 
             //act
-            sut.DownloadVersion("any", "any");
+            sut.DownloadVersion("any", "any", hashFile);
 
             //assert
-            mockEffectiveStrategy.Verify(x => x.DownloadVersion(It.IsAny<string>(), It.IsAny<string>()), Times.Never);
+            mockEffectiveStrategy.Verify(x => x.DownloadVersion(It.IsAny<string>(), It.IsAny<string>(), null), Times.Never);
             mockFileProxy.Verify(x => x.CopyFile(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<bool>()));
         }
 
         [Test]
         public void GetLatestVersion_PaketHashFileExistsButIsCorrupt_DeleteHashFileAndThrowException()
         {
+            const string hashFile = @"C:\hash.txt";
+
             //arrange
             mockEffectiveStrategy.Setup(x => x.GetLatestVersion(true)).Throws<WebException>().Verifiable();
             mockFileProxy.Setup(x => x.OpenRead(It.IsAny<string>())).Returns(new MemoryStream(Guid.NewGuid().ToByteArray()));
-            mockFileProxy.Setup(x => x.ReadAllLines(It.IsAny<string>())).Returns(new[] { Guid.NewGuid().ToString().Replace("-", String.Empty) + " not-paket.exe" });
+            mockFileProxy.Setup(x => x.ReadAllLines(hashFile)).Returns(new[] { Guid.NewGuid().ToString().Replace("-", String.Empty) + " not-paket.exe" });
             mockFileProxy.Setup(x => x.FileExists(It.IsAny<string>())).Returns(true);
 
             //act
-            Assert.Throws<InvalidDataException>(() => sut.DownloadVersion("any", "any"));
+            Assert.Throws<InvalidDataException>(() => sut.DownloadVersion("any", "any", hashFile));
 
             //assert
-            mockEffectiveStrategy.Verify(x => x.DownloadVersion(It.IsAny<string>(), It.IsAny<string>()), Times.Never);
-            mockFileProxy.Verify(x => x.DeleteFile(ItHasFilename("paket-sha256.txt")));
+            mockEffectiveStrategy.Verify(x => x.DownloadVersion(It.IsAny<string>(), It.IsAny<string>(), null), Times.Never);
+            mockFileProxy.Verify(x => x.DeleteFile(hashFile));
         }
 
         public string ItHasFilename(string filename)
         {
             return It.IsRegex($@"\w*[\\/]{filename}");
+        }
+
+        [Test]
+        [TestCase(true)]
+        [TestCase(false)]
+        public void CanDownloadHashFile(bool can)
+        {
+            mockEffectiveStrategy.SetupGet(x => x.CanDownloadHashFile).Returns(can);
+            Assert.That(sut.CanDownloadHashFile, Is.EqualTo(can));
+        }
+
+        [Test]
+        public void DownloadHashFile_NotSupported()
+        {
+            mockEffectiveStrategy.SetupGet(x => x.CanDownloadHashFile).Returns(false);
+
+            var hashFilePath = sut.DownloadHashFile("42.0");
+
+            Assert.That(hashFilePath, Is.Null);
+        }
+
+        [Test]
+        public void DownloadHashFile_PresentInCache()
+        {
+            mockEffectiveStrategy.SetupGet(x => x.CanDownloadHashFile).Returns(true);
+            var pathInCache = sut.GetHashFilePathInCache("42.0");
+            mockFileProxy.Setup(x => x.FileExists(pathInCache)).Returns(true);
+
+            var hashFilePath = sut.DownloadHashFile("42.0");
+
+            Assert.That(hashFilePath, Is.EqualTo(pathInCache));
+            mockEffectiveStrategy.Verify(x => x.DownloadHashFile(It.IsAny<string>()), Times.Never);
+        }
+
+        [Test]
+        public void DownloadHashFile_NotInCache()
+        {
+            const string pathFromEffectiveStrategy = @"C:\hash.txt";
+            mockEffectiveStrategy.Setup(x => x.DownloadHashFile(It.IsAny<string>())).Returns(pathFromEffectiveStrategy);
+            mockEffectiveStrategy.SetupGet(x => x.CanDownloadHashFile).Returns(true);
+            var pathInCache = sut.GetHashFilePathInCache("42.0");
+            mockFileProxy.Setup(x => x.FileExists(pathInCache)).Returns(false);
+
+            var hashFilePath = sut.DownloadHashFile("42.0");
+
+            Assert.That(hashFilePath, Is.EqualTo(pathInCache));
+            mockEffectiveStrategy.Verify(x => x.DownloadHashFile("42.0"), Times.Once);
+            mockFileProxy.Verify(x => x.CopyFile(pathFromEffectiveStrategy, pathInCache, true), Times.Once);
         }
     }
 }